+++ ChangeLog.md
@@ -13,6 +13,9 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 
 # Feature Modifications
 
-1. Ported rule `blockStateTextureSync`.
+## Lifetime Tracker
+
+1. Add rule `lifetimeTrackBySize`.
+2. Chicken jockeys and spider jockeys will be tracked.
 
 # Code Details